requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-admin==2.0.2
django-excel-response2==3.0.6
django-ninja==1.1.0
django-six==1.0.5
excel-base==1.0.4
idna==3.7
isoweek==1.3.3
lxml==5.2.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
plotly==5.22.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
screen==1.0.1
six==1.16.0
sqlparse==0.5.0
tenacity==8.3.0
TimeConvert==3.0.13
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
xlwt==1.3.0
